Added WinRT toolkit and GridSplitter
Added a GridSplitter to the screen to host the list of locations. Some
location names are really long, and this lets the user show/hide it.
There is a bug where if you drag the GridSplitter all the way to the
right causing the map to be tiny. This causes the app to crash.
I worked around it by putting a MaxWidth on the list so it can't be
dragged too far to the right. I also added a MinWidth to the Map grid,
to stop it becoming too small when the orientation changes.